// app/utils/interfaces.ts
export interface DBConnectionOptions {
    host: string;
    user: string;
    password: string;
    database: string;
  }

export interface intro {
  title: string;
  name: string;
  profile: string;
  designation: string;
  description: string;
  year_of_experience: Number;
  project_completed: Number;
}